# move sector cache config out of repetition loop