--- a/csrspmv.c
+++ b/csrspmv.c
@@ -1130,11 +1130,6 @@ static int csrgemv(
     const double * __restrict a)
 {
 #if defined(__FCC_version__) && defined(USE_A64FX_SECTOR_CACHE)
-#ifdef A64FX_SECTOR_CACHE_L1_WAYS
-    #pragma procedure scache_isolate_way L2=A64FX_SECTOR_CACHE_L2_WAYS L1=A64FX_SECTOR_CACHE_L1_WAYS
-#else
-    #pragma procedure scache_isolate_way L2=A64FX_SECTOR_CACHE_L2_WAYS
-#endif
     #pragma procedure scache_isolate_assign a, colidx
 #endif
 
@@ -1164,11 +1159,6 @@ static int csrgemvsd(
     const double * __restrict ad)
 {
 #if defined(__FCC_version__) && defined(USE_A64FX_SECTOR_CACHE)
-#ifdef A64FX_SECTOR_CACHE_L1_WAYS
-    #pragma procedure scache_isolate_way L2=A64FX_SECTOR_CACHE_L2_WAYS L1=A64FX_SECTOR_CACHE_L1_WAYS
-#else
-    #pragma procedure scache_isolate_way L2=A64FX_SECTOR_CACHE_L2_WAYS
-#endif
     #pragma procedure scache_isolate_assign a, ad, colidx
 #endif
 
@@ -2173,6 +2163,14 @@ int main(int argc, char *argv[])
     }
 #endif
 
+#if defined(__FCC_version__) && defined(USE_A64FX_SECTOR_CACHE)
+#ifdef A64FX_SECTOR_CACHE_L1_WAYS
+    #pragma statement scache_isolate_way L2=A64FX_SECTOR_CACHE_L2_WAYS L1=A64FX_SECTOR_CACHE_L1_WAYS
+#else
+    #pragma statement scache_isolate_way L2=A64FX_SECTOR_CACHE_L2_WAYS
+#endif
+#endif
+
     /* 5. compute the matrix-vector multiplication. */
 #ifdef _OPENMP
     #pragma omp parallel
@@ -2234,6 +2232,10 @@ int main(int argc, char *argv[])
                     (double) max_bytes * 1e-9 / (double) timespec_duration(t0, t1));
         }
     }
+
+#if defined(__FCC_version__) && defined(USE_A64FX_SECTOR_CACHE)
+    #pragma statement end_scache_isolate_way
+#endif
 
 #ifdef HAVE_PAPI
     if (papi_opt.event_file) {