This moves the sector cache config out of repetition loop