@jamtrott 2023-05-05
Hi Sergej,

Great! This removed the overhead from enabling/disabling the sector cache.

The table below illustrates the difference. The numbers in the first column is the performance with sector cache disabled. The second and third columns show the results with sector cache enabled (with 4 L2 ways for non-reusable data).

```
                     No Sector   Sector        Sector
                     Cache       Cache (old)   Cache (new)
  Matrix             [Gflop/s]   [Gflop/s]     [Gflop/s]
 ------------------ ----------- ------------- -------------
  Koutsovasilis/F1      33.905        42.600        52.687  
  AG-Monien/debr        18.172        12.371        18.134
```

To get the numbers above, I compiled the code on the Wisteria cluster with the following commands:
```
$ module load fj
$ make CC=fccpx CFLAGS="-Kfast -Khpctag -Kocl -Kopenmp -DHAVE_ALIGNED_ALLOC -DHAVE_LIBZ -DHAVE_PAPI -DUSE_A64FX_SECTOR_CACHE -DA64FX_SECTOR_CACHE_L2_WAYS=4" LDFLAGS="-lpapi -lz"
```
I then ran the code as follows:
```
$ pjsub --interact -g jh180024o -L rscgrp=interactive-o -X
$ OMP_DISPLAY_ENV=true OMP_NUM_THREADS=48 OMP_PROC_BIND=close FLIB_HPCFUNC=TRUE FLIB_SCCR_CNTL=TRUE FLIB_L1_SCCR_CNTL=FALSE XOS_MMM_L_HPAGE_TYPE=hugetlbfs XOS_MMM_L_PAGING_POLICY=demand:demand:demand ./csrspmv --verbose -q --repeat=100 -z ../../mtx/original/Koutsovasilis/F1.mtx.gz
mtxfile_read: 15.195427 seconds (26.6 MB/s)
csr_from_coo: 0.472682 seconds, 343,791 rows, 343,791 columns, 26,837,113 nonzeros, 24 to 435 nonzeros per row, 48 threads, 7,162 to 7,163 rows per thread, 417,845 to 1,373,845 nonzeros per thread
gemv: 0.001136 seconds (11.968 Gnz/s, 47.268 Gflop/s, 290.9 to 477.5 GB/s)
gemv: 0.001021 seconds (13.304 Gnz/s, 52.545 Gflop/s, 323.3 to 530.8 GB/s)
gemv: 0.001026 seconds (13.250 Gnz/s, 52.331 Gflop/s, 322.0 to 528.7 GB/s)
gemv: 0.001023 seconds (13.288 Gnz/s, 52.479 Gflop/s, 322.9 to 530.2 GB/s)
gemv: 0.001073 seconds (12.669 Gnz/s, 50.036 Gflop/s, 307.9 to 505.5 GB/s)
gemv: 0.001021 seconds (13.310 Gnz/s, 52.566 Gflop/s, 323.5 to 531.0 GB/s)
[...]
```